mapping: don't attempt to use a property we don't know about
We can end up with an error (e.g.:
<class 'pysqlite2.dbapi2.OperationalError'>: no such column: flags
) if the dataId contains properties (keywords) other than ones we
know are in the registry.  In particular, we need to filter out the
'flags' that are passed in following #2940 (for afw tables, not
reading footprints); these are treated as part of the dataId.

I hope this will be resolved in a better way as part of the butler
overhaul.