Change butler queryMetadata test so lsstSim is not needed
ticket1640 was testing whether **kwargs worked in a queryMetadata
call in addition to explicit keyword=value arguments when the ordering
of the keywords differed.

The test required lsstSim data because that was how the problem was
reported in the original ticket. Now use the test mapper from
tests/cameraMapper.py. Note that because we can not guarantee how
dicts are unpacked or what happens to keyword argument ordering
this test could in theory fail on one system and pass on another.

The original fix for this was in c25a6bde54b60e3793c9b9875543c97b47f1ff14

This test does not in fact trigger the failure reported in the
original ticket. The failure relied critically on the dict with
keys "sensor", "visit" and "raft" in dataId being returned in
the same order as a new dict. A test with those three query items
would require changes to the test data files. This is deemed to
be a low priority given the new butler implementation.