CameraMapper: move camera construction to a private method
Add CameraMapper._makeCamera to make the camera. The default
implementation works just like the code that was in the constructor
except for one change: it raises an exception if the necessary
information is not available. This reduces the odds of a silent
failure, but required adding data for a small camera to the
test repository and its paf files.