fix broken getCameraName implementaiton
this seems very inelegant but works I think.
Sometimes the name can be e.g. __main__.CameraMapper
and sometimes it can be cameraMapper.CameraMapper;
It depends on how the script is run.
I think always the name will be dot-qualified, and
we always want the class name, which is after the
dot. So: find the dot, strip everything before it
and then perform the globbing as before.