Add MakeRawVisitInfo
Add MakeRawVisitInfo, a base class functor to make
VisitInfo from the FITS header of a raw image
and remove the metadata keywords that were used.
The base class handles `exposureTime` and `date` if specified
by one common set of FITS kewyords, but individual cameras
will almost always want to override this class.

Add a test for the new class.