CameraMapper: simplify setting of dataset mappings
Extra dataset features (e.g., "raw_sub", "raw_md", "raw_filename") are
are created from the original mapping (e.g., "raw") by defining a bunch
of extra methods. This patch replaces the sea of
    if not hasattr(self, someName):
        setattr(self, someName, someImpl)
calls with a common function to set the implementations for each extra
dataset feature. This clarifies the code and makes it easier to add more
extra features.

Moved the setting of dataset mappings into its own method, as it's
specialised and too much to have in __init__.