Add yamlCamera.py to obs_base
The initial commit of the yaml camera description to obs_base.
Those who want to make a yaml-based camera description, it can be imported with
from lsst.obs.base.yamlCamera import YamlCamera
and an example implementation can be found in
obs_comCam/python/lsst/obs/comCam/comCam.py and
obs_comCam/policy/camera.yaml
though it should be noted that this is not the most simple of starting examples
on which to base other camera descriptions due to the layout of LSST chips, and
this being for a whole raft.