Get FITS image write options into the persistence framework
Introduces the concept of recipes for writing. The policy/writeRecipes.yaml
file has the recipes, allowing recipes to be selected by symbolic name.
Supplemental recipes can be specified in the obs package. We include recipes
for no image compression (the default), lossless and lossy compression
(included for convenience because I hope to evaluate it and activate
it soon).