Change storage type to ParquetStorage
This changes the storage type of deepCoadd_obj to ParquetStorage,
as well as the python type.  Notably, this dataset (like all
ParquetStorage datasets) should be written as a pandas DataFrame,
even though it gets loaded as a (Multilevel)ParquetTable, which is
an object that allows lazy loading of the data, as needed.