Make IsrTask a valid CmdLineTask.
The IsrTask was not usable as a CmdLineTask, as it was written to be
called as a subtask of ProcessCcdTask.  This allows the default
IsrTask to be retargeted to camera-specific versions of the ISR
processing.  A new RunIsrTask now exists to hold IsrTask for
retargetting in standalone processing, and both processCcd.py and
runIsr.py now read from the same configuration files for each obs_
package.  This has involved migrating configuration values from
config/processCcd.py to config/isr.py for many obs_ packages.