Change creation pattern and support distortion in YamlCamera
Provide a makeCamera function rather than inheriting from Camera.

The support for distortion models is mostly in makeTransformDict,
and required clarification that all models are
	*from* native *to* other
which is a change from at least some other cameras.

N.b. makeCamera required that we changed a number of methods to be
functions, which accounts for most of the changes.