Implement gen3 fitsRaw detector and wcs readers
This updates the existing `makeWcs` to attempt to create a SkyWcs from
the VisitInfo and Detector objects, to match the gen2 code.

Note: this logic is more strict than the gen2 version: if we decide we need to
have less harsh failure modes (e.g. mangled FITS headers, or calibs), we can
sort that out once we have identified the use cases.

Add unittests of wcs creation, covering a few different failure modes.