Rewrite Gen3 raw ingest task.
In addition to updating to daf.butler's new dimensions and query APIs,
this removes several bells and whistles from the configuration options
dealing with transactions and restarts.  It's not clear we had use
cases for all of those options, and some of them created serious
implementation challenges in the butler.  Going forward, we'll add
those kinds of options either because we have a clear use case or
we're certain the implementation is straightforward.

This also structures things so the ingest task can be used to help
convert Gen2 repos to Gen3, so we can stop maintaining a separate path
for extracting observational dimensions from raws there.