Convert subfilter and numSubfilter types to str
This is necessary for compatibility with coadd source detection and measurements using multiband.py . The datasetType in the argument parser is hardcoded as deepCoadd, which lacks these two keys. Since the argument parser defaults to type str for any keys that it finds that are not present in the definitions set by the datasetType, we can either change these keys to match the default type, or define wrappers around the tasks in multiband.py to replace the hardcoded defaults with dcrCoadd. Since this problem is likely to go away with the Gen 3 butler, it seems best to simply change the type. It should be noted that these keys both already had type str in `obs_decam`'s DecamMapper.yaml.