Add default root-repo skymap support in gen2to3 for brightObjectMasks.
It's common for Gen2 repos to have datasets that use skymap-based data
IDs in the root repo even though there is no skymap dataset in the
root repo (primarily bright object masks).

This change makes it possible to convert those ill-formed repos to
Gen3 if the user tells us (via configuration) which skymap should be
assumed.