makeRawVisitInfo: don't strip header by default
RFC-576 established that we would propagate all input headers to the
output files, but we have been stripping header keywords used in the
creation of the VisitInfo. Made this feature optional (through a
parameter provided to the Ctor), and turned it off by default, so that
we won't lose the header keywords.

The same is done for makeRawVisitInfoViaObsInfo.