Remove check for visit IDs being disjoint from exposure IDs.
I'd like to reserve exposure ID values for use by the one-to-one
grouping algorithm, in case we want to use that for LSST data, too,
but this is already inconsistent with the visit IDs we get from
ObservationInfo, and I don't want to hold up this ticket deciding what
to do about that.

Ultimately we may want to use instrument's visit_max and/or
exposure_max (and move the former to visit_system?) to enshrine the
ranges reserved for different grouping algorithms someday.