remove unneeded command name
The subcommand name comes from the function name, with the underscore
changed to a dash.

If named, the subcommand name should have had a dash ("register-instrument"),
but it didn't matter so much because Click was changing the underscore
to a dash with the explicitly provided command name anyway. But this
is more correct.