Refactor data ID translator construction.
This mostly moves code around, defining a few new interfaces while
keeping logic unchanged.  It'd have been nice to split it up into
multiple commits, but I didn't see a way to do that without breaking
things in the interim.  So here's a list of what's moved where, and
why:

 - The "rules" for data ID translation have been moved from global
   state (a class attribute of the Translator class) to per-instance
   state on the new TranslatorFactory class.  Same with the
   makeMatching method.

 - The mostly-instrument-generic rule definitions have been moved from
   the command-line script module to a new addGenericInstrumentRules
   method on TranslatorFactory.

 - Actually constructing a TranslatorFactory is now something
   ConvertRepoTask does, delegating that to the Instrument classes,
   where there is a new abstract method.

 - The concrete implementations of that method in obs_subaru,
   obs_decam, and obs_lsst call addGenericInstrumentRules with the
   appropriate arguments for that instrument, and in the case of HSC
   and DECam then add less-generic rules that were moved from
   import-time code (for HSC) and the config override file (for DECam).

In addition to just avoiding global state, this makes it possible once
again to invoke ConvertRepoTask directly for more advanced usage than
is possible with the command-line script (as we do in HSC RC2
bootstrapping).