Change ingest-raws so that it derives collection name from raw data
The output_run is now optional on the command line and if not
specified a suitable value will be used that is derived from
the data. We now allow files from multiple instruments
to be ingested at once (assuming they are compatible with
the Task being used) into their own collections.