Change gen2to3 so that it asks the mapper for the gen3 instrument
The gen2 mapper associated with a gen2 repo maps to a single
gen3 instrument. Rather than asking the caller to specify the
gen3 instrument explicitly, ask the mapper to tell us.

This removes the config entry for instrument and the command
line override.

It does currently require an entire Butler instance to be
created solely to read the mapper class. This may be
slow.