Add ability to persist plots to a sub-directory
There is an option in the pipe_analysis scripts for the user to add
a "subdir" to the plot output path.  This is useful in testing when
not wanting to overwrite the output files in a given rerun for each
test run (e.g. for comparison purposes).  This adds the appropriate
field in the dataset template definitions.