Only run full raw read in one of the ingest tests
We do not need to read the same raws in every ingest check.
We also do not need to read every raw dataset that was ingested.
Read the first raw completely but for subsequent raws
just read some metadata to demonstrate that the datastore
can read them.

Changing this can make significant improvements to test
runtimes.