Attach afw Reader instance as Formatter instance attribute.
This will soon be important for allowing derived-class method overrides
access to the Reader object.  I'm a little uncomfortable with this,
but it seems to be the current pattern for this sort of thing in
Formatter objects, and I don't a near-term easier way.  Hopefully this
will be at least mitigated by DM-28698, e.g. by breaking the
inheritance relationship with FitsRawFormatterBase, which doesn't use
Reader classes at all, AFAICt.