Remove special metadata logic from non-raw Exposure formatter.
Now that the raw Exposure Formatter no longer depends on this one, we
can stop patching the headers (since they'll have already been patched
when we read the raws), and we can remove this version of the stripping
logic and just delegate to afw's ExposureFitsReader implementation of
that.

This revealed a small problem in some test code that was trying to get
compression-related headers that we'd normally want to strip (the new
stripping is more aggressive than the old one), but the fix is easy.