Remove unused readRawHeaderWcs.
This seems to have been added to support a Gen3 equivalent of Gen2's
raw_header_wcs dataset.  But there is no such equivalent in Gen3, and
no one seems to have noticed its absence.  This is probably because
Gen3 has always been consistent about using a boresight+cameraGeom WCS
for raws, and debugging when that happened (or didn't) was what made
the Gen2 dataset useful.