Set defineVisits padding to 250 pixels
This value is the default chosen based on HSC data.  It may need adjusting
for other cameras (if, e.g. they suffer significant pointing errors or don't
include a good distortion model such that the "raw" WCS is not a good
representation of the "true" WCS).  Regardless, the default of zero was
certainly too small (i.e. far from satisfying the "conservatively large"
conjecture in the doc), so this has been deemed a best guess default for
now.