Fix newly-revealed preexisting bug in defineVisits test.
I'm not entirely sure how this worked before - the test uses "raw_dict"
as the dataset type name for raws, while calling a script that
hard-coded "raw" for that name.  But daf_butler changes on this ticket
caused the ticket to fail, and this change was always the right fix.