Change import style to ease formatting.
Black will reformat this long lambda to extend beyond our line length,
with then upsets flake8.  I can't really blame it, as anything (like
the current formatting) that does fit within our line length is pretty
hard to parse.

Anyhow, changing this one import is enough to make it fit, so I'm
punting on the more general problem of how to make black and flake8
play nicely on this issue.