Remove internal multiprocessing from DefineVisitsTask.
This had stopped working a while ago, if it ever worked at all, and
with us now creating raw WCSs from Registry's DimensionRecords instead
of reading raw files, it's not worth trying to get working.

Now that we don't have to use "map" or its multiprocessing
counterparts, we can also simplify the implementation by removing a
few methods whose only job was to make a call suitable for "map".