Add/fix type annotations in Instrument.
This is mostly adding missing annotations and simple fixes like
accounting for possible None return values.  It also involves renaming
some variables because MyPy only allows local variable types to
narrow, not widen.  The new names are a bit clearer anyway ("cls"
should not be used for the _name_ of a type!).