Add/fix type annotations in ExposureAssembler.
There was a lot of inconsistency between the signatures here and those
of the base classes in daf_butler, and resolving them required some
changes in daf_butler as well.  Some of these involved optional
arguments in the base class that simply were not present in
ExposureAssembler, which I've handled by adding them but raising
NotImplementedError when not-None values are passed.