ingest: make customisation easier
When instruments need additional exposure properties and/or
dimensions, they need a way to provide that information to the
RawIngestTask. We already have the ability to subclass the
RawIngestTask (via the --ingest-task command-line argument),
but there was no way for the subtask to specify the additional
information.

Moved some defaults (the ObservationInfo class and the list of
required ObservationInfo fields) into class variables, and added
methods to provide the required registry records. An important
conceptual addition is the idea of 'extra records': records
which are added to the registry in addition to the exposure
record. These can be used to satisfy foreign key constraints
when the exposure relates to dimensions beyond the standard set.