Allow multi-snap visits to be redefined incrementally
Previously the multi-snap visit was only defined when the second
snap arrived. Now we can define the multi-snap visit even if
we only have one snap (so we get two visits defined). When the
second snap arrives there will be three visits and the multi-snap
visit will be updated if --update-records is enabled.