Expand dataset refs before passing them to DefineVisitsTask (DM-38678)
DefineVisitsTask calls `expandDataId()` internaly to expand each dataset ref
in a loop, but this is not a recommended way. This patch adds `expanded()`
call on the iterator before passing it to the task to achieve the same. I have
not removed `expandDataId()` from the task, as it may be also be used in other
contexts which use un-expanded refs.