First draft of a simplified simulation-ingest tool.
This is hanging out here outside the Python module because I wanted to
start it as a standalone module that could be copied around anywhere
and used with a released version of obs_base (I believe the first
consumers are conda stackvana users).  After it's stabilized I'll move
it into the package.

This utility goes directly from an lsst.afw.image.Exposure to visit
and visit_detector_region dimension records.  This seems way simpler
than going through ObservationInfo, exposure dimension records, and
DefineVisitsTask, but that may be enough more general (it'd let this
handle postISRCCD simulations) that it may be worthwhile, so I still
plan to try that.

This makes no attempt to be useful for simulated coadds (coadds are
actually sufficiently easier that I'm not sure there would be anything
nontrivial in common).